move globals into state/ directory
