+++ src/cursor/cursor.c
@@ -2,8 +2,8 @@
 #include "extensions/time.h"
 #include "output/output.h"
 #include "output/outputs.h"
-#include "scene.h"
-#include "seat.h"
+#include "state/scene.h"
+#include "state/seat.h"
 
 struct wlr_cursor* catnip_cursor = NULL;
 struct wlr_xcursor_manager* catnip_xcursor_manager = NULL;
