+++ src/cursor/init.c
@@ -1,11 +1,11 @@
 #include "init.h"
-#include "backend.h"
 #include "config.h"
 #include "cursor/cursor.h"
 #include "cursor/lua_cursor.h"
 #include "extensions/wayland.h"
 #include "output/output_layout.h"
-#include "seat.h"
+#include "state/backend.h"
+#include "state/seat.h"
 #include <string.h>
 
 static struct wl_listener cursor_motion_listener;
