+++ src/keyboard/key_event.c
@@ -1,5 +1,5 @@
 #include "key_event.h"
-#include "backend.h"
+#include "state/backend.h"
 #include <xkbcommon/xkbcommon-keysyms.h>
 
 static bool
