+++ src/lua_catnip.c
@@ -2,12 +2,12 @@
 #include "canvas/lua_canvas.h"
 #include "config.h"
 #include "cursor/lua_cursor.h"
-#include "display.h"
 #include "extensions/string.h"
 #include "keyboard/lua_keyboard_list.h"
 #include "lua_events.h"
 #include "output/lua_output_list.h"
-#include "seat.h"
+#include "state/display.h"
+#include "state/seat.h"
 #include "window/lua_window_list.h"
 #include "window/window.h"
 #include "window/windows.h"
