+++ src/main.c
@@ -1,19 +1,19 @@
-#include "allocator.h"
-#include "backend.h"
 #include "cli.h"
 #include "config.h"
 #include "cursor/init.h"
-#include "display.h"
-#include "event_loop.h"
 #include "keyboard/init.h"
 #include "lua_events.h"
 #include "output/init.h"
-#include "renderer.h"
-#include "scene.h"
-#include "seat.h"
+#include "state/allocator.h"
+#include "state/backend.h"
+#include "state/display.h"
+#include "state/event_loop.h"
+#include "state/renderer.h"
+#include "state/scene.h"
+#include "state/seat.h"
+#include "state/xdg_shell.h"
 #include "utils/log.h"
 #include "window/init.h"
-#include "xdg_shell.h"
 #include <stdlib.h>
 #include <wlr/types/wlr_data_device.h>
 #include <wlr/types/wlr_subcompositor.h>
